refactor(fee): moving calculate_tx_gas_usage_vector to ActualCostBuilder
