+++ crates/blockifier/src/fee/gas_usage_test.rs
@@ -1,23 +1,14 @@
 use pretty_assertions::assert_eq;
 use rstest::{fixture, rstest};
 use starknet_api::hash::StarkFelt;
-use starknet_api::stark_felt;
-use starknet_api::transaction::{EventContent, EventData, EventKey, L2ToL1Payload};
+use starknet_api::transaction::{EventContent, EventData, EventKey};
 
-use crate::execution::call_info::{
-    CallExecution, CallInfo, MessageToL1, OrderedEvent, OrderedL2ToL1Message,
-};
+use crate::execution::call_info::{CallExecution, CallInfo, OrderedEvent};
 use crate::fee::eth_gas_constants;
-use crate::fee::gas_usage::{
-    calculate_tx_gas_usage_vector, get_consumed_message_to_l2_emissions_cost, get_da_gas_cost,
-    get_log_message_to_l1_emissions_cost, get_message_segment_length, get_tx_events_gas_cost,
-};
+use crate::fee::gas_usage::{get_da_gas_cost, get_tx_events_gas_cost};
 use crate::state::cached_state::StateChangesCount;
-use crate::test_utils::contracts::FeatureContract;
-use crate::test_utils::CairoVersion;
 use crate::transaction::objects::GasVector;
-use crate::transaction::test_utils::calculate_class_info_for_testing;
-use crate::utils::{u128_from_usize, usize_from_u128};
+use crate::utils::u128_from_usize;
 use crate::versioned_constants::VersionedConstants;
 
 #[fixture]
@@ -120,268 +111,6 @@ fn test_get_da_gas_cost_basic(#[case] state_changes_count: StateChangesCount) {
         GasVector { l1_gas: 0, l1_data_gas: u128_from_usize(manual_blob_gas_usage).unwrap() },
         computed_gas_vector
     );
-}
-
-/// This test goes over seven cases. In each case, we calculate the gas usage given the parameters.
-/// We then perform the same calculation manually, each time using only the relevant parameters.
-/// The seven cases are:
-///     1. An empty transaction.
-///     2. A Declare transaction.
-///     3. A DeployAccount transaction.
-///     4. An L1 handler.
-///     5. A transaction with L2-to-L1 messages.
-///     6. A transaction that modifies the storage.
-///     7. A combination of cases 4. 5. and 6.
-// TODO(Aner, 29/01/24) Refactor with assert on GasVector objects.
-// TODO(Aner, 29/01/24) Refactor to replace match with if when formatting is nicer
-#[rstest]
-fn test_calculate_tx_gas_usage_basic(#[values(false, true)] use_kzg_da: bool) {
-    // An empty transaction (a theoretical case for sanity check).
-    let versioned_constants = VersionedConstants::default();
-    let empty_tx_gas_usage_vector = calculate_tx_gas_usage_vector(
-        &versioned_constants,
-        std::iter::empty(),
-        StateChangesCount::default(),
-        0,
-        0,
-        None,
-        None,
-        use_kzg_da,
-    )
-    .unwrap();
-    assert_eq!(empty_tx_gas_usage_vector, GasVector { l1_gas: 0, l1_data_gas: 0 });
-
-    // Declare.
-    for cairo_version in [CairoVersion::Cairo0, CairoVersion::Cairo1] {
-        let empty_contract = FeatureContract::Empty(cairo_version).get_class();
-        let class_info = calculate_class_info_for_testing(cairo_version, empty_contract);
-        let code_milligas_cost = u128_from_usize(
-            (class_info.bytecode_length() + class_info.sierra_program_length)
-                * eth_gas_constants::WORD_WIDTH
-                + class_info.abi_length,
-        )
-        .unwrap()
-            * versioned_constants.l2_resource_gas_costs.milligas_per_code_byte;
-        let manual_gas_vector =
-            GasVector { l1_gas: code_milligas_cost / 1000, ..Default::default() };
-        let declare_gas_usage_vector = calculate_tx_gas_usage_vector(
-            &versioned_constants,
-            std::iter::empty(),
-            StateChangesCount::default(),
-            0,
-            0,
-            None,
-            Some(class_info),
-            use_kzg_da,
-        )
-        .unwrap();
-        assert_eq!(manual_gas_vector, declare_gas_usage_vector);
-    }
-
-    // DeployAccount.
-
-    let deploy_account_state_changes_count = StateChangesCount {
-        n_storage_updates: 0,
-        n_class_hash_updates: 1,
-        n_compiled_class_hash_updates: 0,
-        n_modified_contracts: 1,
-    };
-
-    // Manual calculation.
-    let calldata_length = 0;
-    let signature_length = 2;
-    let calldata_and_signature_milligas_cost = u128_from_usize(calldata_length + signature_length)
-        .unwrap()
-        * versioned_constants.l2_resource_gas_costs.milligas_per_data_felt;
-    let manual_starknet_gas_usage = calldata_and_signature_milligas_cost / 1000;
-    let manual_gas_vector = GasVector { l1_gas: manual_starknet_gas_usage, ..Default::default() }
-        + get_da_gas_cost(deploy_account_state_changes_count, use_kzg_da);
-
-    let deploy_account_gas_usage_vector = calculate_tx_gas_usage_vector(
-        &versioned_constants,
-        std::iter::empty(),
-        deploy_account_state_changes_count,
-        calldata_length,
-        signature_length,
-        None,
-        None,
-        use_kzg_da,
-    )
-    .unwrap();
-    assert_eq!(manual_gas_vector, deploy_account_gas_usage_vector);
-
-    // L1 handler.
-
-    let l1_handler_payload_size = 4;
-    let l1_handler_gas_usage_vector = calculate_tx_gas_usage_vector(
-        &versioned_constants,
-        std::iter::empty(),
-        StateChangesCount::default(),
-        l1_handler_payload_size,
-        signature_length,
-        Some(l1_handler_payload_size),
-        None,
-        use_kzg_da,
-    )
-    .unwrap();
-
-    // Manual calculation.
-    let message_segment_length = get_message_segment_length(&[], Some(l1_handler_payload_size));
-    let calldata_and_signature_milligas_cost =
-        u128_from_usize(l1_handler_payload_size + signature_length).unwrap()
-            * versioned_constants.l2_resource_gas_costs.milligas_per_data_felt;
-    let manual_starknet_gas_usage = message_segment_length * eth_gas_constants::GAS_PER_MEMORY_WORD
-        + eth_gas_constants::GAS_PER_COUNTER_DECREASE
-        + usize_from_u128(
-            get_consumed_message_to_l2_emissions_cost(Some(l1_handler_payload_size)).l1_gas,
-        )
-        .unwrap()
-        + usize_from_u128(calldata_and_signature_milligas_cost / 1000).unwrap();
-    let manual_sharp_gas_usage =
-        message_segment_length * eth_gas_constants::SHARP_GAS_PER_MEMORY_WORD;
-    let manual_gas_computation = GasVector {
-        l1_gas: u128_from_usize(manual_starknet_gas_usage + manual_sharp_gas_usage).unwrap(),
-        l1_data_gas: 0,
-    };
-
-    assert_eq!(l1_handler_gas_usage_vector, manual_gas_computation);
-
-    // Any transaction with L2-to-L1 messages.
-
-    let mut call_infos = Vec::new();
-
-    for i in 0..4 {
-        let payload_vec = vec![stark_felt!(0_u16); i];
-
-        let call_info = CallInfo {
-            execution: CallExecution {
-                l2_to_l1_messages: vec![OrderedL2ToL1Message {
-                    message: MessageToL1 {
-                        payload: L2ToL1Payload(payload_vec),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                }],
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-
-        call_infos.push(call_info);
-    }
-
-    // l2_to_l1_payload_lengths is [0, 1, 2, 3]
-    let call_infos_iter = call_infos.iter();
-    let l2_to_l1_payload_lengths: Vec<usize> = call_infos_iter
-        .clone()
-        .flat_map(|call_info| call_info.get_sorted_l2_to_l1_payload_lengths().unwrap())
-        .collect();
-
-    let l2_to_l1_state_changes_count = StateChangesCount {
-        n_storage_updates: 0,
-        n_class_hash_updates: 0,
-        n_compiled_class_hash_updates: 0,
-        n_modified_contracts: 1,
-    };
-    let l2_to_l1_messages_gas_usage_vector = calculate_tx_gas_usage_vector(
-        &versioned_constants,
-        call_infos_iter.clone(),
-        l2_to_l1_state_changes_count,
-        0,
-        0,
-        None,
-        None,
-        use_kzg_da,
-    )
-    .unwrap();
-
-    // Manual calculation.
-    let message_segment_length = get_message_segment_length(&l2_to_l1_payload_lengths, None);
-    let n_l2_to_l1_messages = l2_to_l1_payload_lengths.len();
-    let manual_starknet_gas_usage = message_segment_length * eth_gas_constants::GAS_PER_MEMORY_WORD
-        + n_l2_to_l1_messages * eth_gas_constants::GAS_PER_ZERO_TO_NONZERO_STORAGE_SET
-        + usize_from_u128(get_log_message_to_l1_emissions_cost(&l2_to_l1_payload_lengths).l1_gas)
-            .unwrap();
-    let manual_sharp_gas_usage = message_segment_length
-        * eth_gas_constants::SHARP_GAS_PER_MEMORY_WORD
-        + usize_from_u128(get_da_gas_cost(l2_to_l1_state_changes_count, use_kzg_da).l1_gas)
-            .unwrap();
-    let manual_sharp_blob_gas_usage =
-        get_da_gas_cost(l2_to_l1_state_changes_count, use_kzg_da).l1_data_gas;
-    let manual_gas_computation = GasVector {
-        l1_gas: u128_from_usize(manual_starknet_gas_usage + manual_sharp_gas_usage).unwrap(),
-        l1_data_gas: manual_sharp_blob_gas_usage,
-    };
-
-    assert_eq!(l2_to_l1_messages_gas_usage_vector, manual_gas_computation);
-
-    // Any calculation with storage writings.
-
-    let n_modified_contracts = 7;
-    let n_storage_updates = 11;
-    let storage_writes_state_changes_count = StateChangesCount {
-        n_storage_updates,
-        n_class_hash_updates: 0,
-        n_compiled_class_hash_updates: 0,
-        n_modified_contracts,
-    };
-    let storage_writings_gas_usage_vector = calculate_tx_gas_usage_vector(
-        &versioned_constants,
-        std::iter::empty(),
-        storage_writes_state_changes_count,
-        0,
-        0,
-        None,
-        None,
-        use_kzg_da,
-    )
-    .unwrap();
-
-    // Manual calculation.
-    let manual_gas_computation = get_da_gas_cost(storage_writes_state_changes_count, use_kzg_da);
-
-    assert_eq!(manual_gas_computation, storage_writings_gas_usage_vector);
-
-    // Combined case of an L1 handler, L2-to-L1 messages and storage writes.
-    let combined_state_changes_count = StateChangesCount {
-        n_storage_updates: storage_writes_state_changes_count.n_storage_updates,
-        n_class_hash_updates: 0,
-        n_compiled_class_hash_updates: 0,
-        n_modified_contracts: storage_writes_state_changes_count.n_modified_contracts
-            + l2_to_l1_state_changes_count.n_modified_contracts,
-    };
-    let gas_usage_vector = calculate_tx_gas_usage_vector(
-        &versioned_constants,
-        call_infos_iter,
-        combined_state_changes_count,
-        l1_handler_payload_size,
-        signature_length,
-        Some(l1_handler_payload_size),
-        None,
-        use_kzg_da,
-    )
-    .unwrap();
-
-    // Manual calculation.
-    let fee_balance_discount = match use_kzg_da {
-        true => 0,
-        false => {
-            eth_gas_constants::GAS_PER_MEMORY_WORD - eth_gas_constants::get_calldata_word_cost(12)
-        }
-    };
-
-    let expected_gas_vector = GasVector {
-        l1_gas: l1_handler_gas_usage_vector.l1_gas
-        + l2_to_l1_messages_gas_usage_vector.l1_gas
-        + storage_writings_gas_usage_vector.l1_gas
-        // l2_to_l1_messages_gas_usage and storage_writings_gas_usage got a discount each, while
-        // the combined calculation got it once.
-        + u128_from_usize(fee_balance_discount).unwrap(),
-        // Expected blob gas usage is from data availability only.
-        l1_data_gas: get_da_gas_cost(combined_state_changes_count, use_kzg_da).l1_data_gas,
-    };
-
-    assert_eq!(expected_gas_vector, gas_usage_vector);
 }
 
 #[test]
